Allow a full 64k address-space.
So rather than allocating the user-provided program to the address
space of the CPU we allocate 64k and then copy the user's program
to the start of it.

This means if the program is 0xFF bytes long the user can copy part
of it to 0xFF00, jump to it, after modifying, & etc.

This closes #6.