Moved Stack increment higher
This can potentially be turned into a information leak.
If SP = 1023, then it will pass the check before being updated to 1024.
This gives us a OOB, which allows u to set the value of SP to 0-0xffff.
With careful heap control, you can get a pointer.
This can be stored/examined by the program using ret,
as they do not do an upper bound check.